Create `MergeItems`, a smart wrapper around a map
Summary:
I want to get rid of MergeItem, which I think is adding unnecessary indirection into
the logic for creating merged FlowInfo, because I need to add more complexity to handle
narrows-vs-values and I don't want things to get out of hand.

But I don't want to expose all the low-level details of map manipulation (which unlike
the actual merging isn't especially interesting or tricky, it's just line noise) to
the `merge_flow` function.

So before I eliminate `MergeItem` (which I just want to turn into a simple `Vec<Flow>`,
I first want to create `MergeItems`, a smarter wrapper around `SmallMap<Name, MergeItem>`
so that the line noise stays out of the way of the essential logic.

The result is actually pretty nice - the match on an entry always felt pretty fiddly
and low-level in `merge_flow`, so I think the result of this little refactor will not
only be simpler but also more readable.

I'll remove MergeInfo in the next diff.

Reviewed By: grievejia

Differential Revision: D83675228

fbshipit-source-id: 306489553d29961f3847b2d6fc7659476974e07a