Separate narrow vs value
Summary:
This diff separates value bindings from narrows in `FlowInfo`.

At this point `FlowInfo` contains both a `value` and a `narrow`, both of which
are optional but at least one is always provided (I considered expressing this
in the type system but it gets a bit messy for limited benefit, we can always
do that later).

The biggest insight from this diff is that we have to decide, when a name
doesn't have a value at the top of the loop but does have a narrow, what to do
- my solution is to treat the loop recursion as a value because we have to
treat anything that *may* turn out to be a value as if it *is* one.

This is a minimal change to get things to compile and run with the new structure,
I'm not yet actually using narrow-vs-value to make decisions at flow merging time;
at the moment, merging always produces a "value" style binding, even if it is
composed entirely of narrows; we'll need to enhance the merging logic to see
actual benefits.

Reviewed By: grievejia

Differential Revision: D83675226

fbshipit-source-id: d76369db5bad5835c118d86e0dfdd62521c58749