Make FlowValue (partial refactor)
Summary:
I'm almost ready to work on splitting FlowInfo's representation
of types in the flow so that values and narrows are tracked
separately.

One problem I ran into is that there are a lot of not-very-interesting
places that have to be re-wired because they are implicitly assuming
a flow is always a value; if I try to do everything in one diff, the
complicated and interesting changes to flow merging will get lost in
all the line noise.

All of these examples are probably bugs in the case where the flow is
actually a narrow (e.g. IDE go-to-def, Function-definition "previous def"
logic, and special export resolution are all things that shouldn't really
care about a narrow, they are about the value bound in the current scope).

To simplify the big cutover that's coming soon, this diff introduces a
`FlowValue` struct to wrap the `idx` and `style` of `FlowInfo`, and exposes an
API for getting the value that assumes it might be None. That lets us
migrate all the uninteresting code blocks ahead of time.

The only place outside of `Flow` itself where I allow code to couple to
the `value` field directly is flow merging, which has to be rewritten
anyway when we separate values from narrows (and indeed that's the whole point!)

NOTE: I omit doc comments on `FlowValue` for now because they make more
sense to add when we actually give it semantics in the upcoming diff, it
feels weird to try to comment in the existing code.

Reviewed By: grievejia

Differential Revision: D83675230

fbshipit-source-id: 9d64385ee93dacaf0ea1f0abdf7ca2efb9fdf28d