Treat narrow-only Phi as a narrow, not a value
Summary:
This change, which only applies to branching control flow (because a
narrow-only loop phi automatically gets optimized away - that works for loops
but not branching because a loop can't actually change the final type by
narrowing, whereas branching can), detects the case where branches all have the
same "value" type and differ only in their "narrow" types.

When that happens, we put the Phi type in the "narrow" slot of `FlowInfo` rather
than the "value" slot, and we preserve the value. The only visible behavior change
is that now the loop phi gets optimized away when narrowing happens inside nested
branching - prior to this change, we would record the nested phi as a "value"
which would cause the loop phi not to be optimized.

Fixes https://github.com/facebook/pyrefly/issues/528
Fixes https://github.com/facebook/pyrefly/issues/812

Reviewed By: grievejia

Differential Revision: D83754264

fbshipit-source-id: 9fbc161f68be4ac84dd987bb51dc36b7195d4ff8