Move, rename, and update the bug tag on a test case
Summary:
In the preceding diff, I noted that my work optimizing away loop Phi when we
only narrow in the loop doesn't actually work for narrows that occur in nested
control flow, because at the moment any phi that *isn't* optimized away
is treated as a "value" type rather than a "narrow" type.

This test case - which has lived in `cylces` rather than `flow` because the
error message is about cycles - actually covers that scenario. I'm going to
attempt to fix it in the next diff, it's nice to put the test in the correct
place with a good description first.

Reviewed By: grievejia

Differential Revision: D83754263

fbshipit-source-id: c1ca1dcef4b05f61ea43010750f2ca4f938915ff