Fix inconsistent override check on classes that do not declare any base class
Summary:
My earlier diff D83597969 ungated the inconsistent override check for dunder methods, but it revealed another pre-existing issue in the check: we were iterating over the base class list to find if the given attribute name is in a base class, but we failed to take into account that base class list can be empty (either when the class itself doesn't declare a base class, or when the class put something on the base class list but it's too dynamic for us to figure out what the base type is). If that list is empty, we will skip all the attribute lookup logic and conclude that the name is never defined. This is OK for non-dunder attributes but for dunder attributes, we ought to check the `object` class (which only has dunder attributes defined in it) as opposed to skipping the attribute lookups.

This diff fixes the issue by making it such that if base class list is empty, we treat it as a singleton list `object`.

Fixes https://github.com/facebook/pyrefly/issues/1223

Reviewed By: rchen152

Differential Revision: D83708691

fbshipit-source-id: 2f36838dfdc6a10d4b86816aaff619a411049a43