Avoid using synthesized range when solving subscript narrow bindings
Summary: We seems to be faking a slice expression when analyzing subscript narrows, re-using a range from another expr for the synthesized expr. This is OK from CLI type checking perspective because we never report type errors on the synthesized expr, but it ended up overwriting legit type traces (which is keyed on ranges). This diff makes it such that a dummy range is used for the fake expr, and make sure that the dummy ranges never get registered in type traces.

Reviewed By: vladima, stroxler

Differential Revision: D83802959

fbshipit-source-id: abebea9ff7da3bdf7383438423d8bbd2984d7920