Drop test requirement that handles always finish
Summary:
These assertions are actually pretty nice, but at the end of this stack I will break them on purpose, with additional laziness for modules which we reach through imports only.

For now, this diff takes the simple approach of just removing the assertions. It would be nice to add back some assertions, but I'm leaving that for future work.

Reviewed By: migeed-z

Differential Revision: D83860238

fbshipit-source-id: fea982d6c3d3192e1a69e64c982e8391fa7ba135