generate baseline file
Summary:
This diff adds --baseline and --update-baseline command line args, and functionality to generate a baseline if --update-baseline is passed.

For now, passing just --baseline does nothing. The way this is set up right now means it will only work when running `pyrefly check`. If we want this to work in the IDE, we can extend it to be a config file option in the future.

Reviewed By: connernilsen

Differential Revision: D83062550

fbshipit-source-id: 0398f4b34b5aea063beb85af0177b429d7fab1a5