Validate that alias configs inheritance is working well
Summary:
For issue https://github.com/facebook/pyrefly/issues/1228

The behavior here is matching the runtime.  The documentation does say that model_config should be inherited. Wondering if there is an exception for those two keywords.

Other than that, since those are the things we support for model_config (extra and aliases), we can mark the issue as resolved? we could note the point above though on the issue.

Reviewed By: rchen152

Differential Revision: D83857296

fbshipit-source-id: 575c0373cb6a34912cf15b8ac0c6ea817f96638c