Throws Bad Instantation Error when Creating an Instance of an Abstract Class (#1192)
Summary:
In class metadata, we store information for abstract classes similar to what we do for protocols:
1. whether the class is abstract (extends ABC or has metaclass=ABCMeta)
2. if so, the list of its members (which may or may not be abstract)

In class field, we store whether the field is abstract.

On class instantiation, if the class is abstract we look up each member and see if the field returned via MRO is abstract. If so, we raise an error.

Pull Request resolved: https://github.com/facebook/pyrefly/pull/1192
closes https://github.com/facebook/pyrefly/issues/510

Reviewed By: grievejia

Differential Revision: D83477589

Pulled By: yangdanny97

fbshipit-source-id: ee2903273e39bd5dbd28394241a8a247890689cb