Export captured variables
Summary:
As titled. We need to know, for each function, its set of captured variables in pysa's analysis.

While pyrefly does a scope analysis during the binding step, it doesn't store the results (at least, not in the form we would want).
This diff tries to build captured variables from a very simple scope analysis and using binding resolution.
This should be mostly correct for our purpose.

Reviewed By: stroxler

Differential Revision: D83676810

fbshipit-source-id: e924ea808478376e7c10668652f674f17109f083