Fix a crash and verify and edge case
Summary:
We do a lot of work in `Definitions` to handle names that are defined
more than once, and that plays into how `Static` models bindings.

But there's an edge case we have to account for: in `definitions`, we
have special hooks so that mutable captures and delete-only variables
behave differently from other multiple definitions.

We need to make sure that when function parameters - which we can think
of as assignments that live above the body - are folded together with
the definitions we preserve this behavior.

This diff
- verifies that mutable captures behave as intended (we complain that the
  mutable capture is invalid - interestingly this coder crashes the python
  compiler *without an error message* so it's nice that we can do better than
  CPython here)
- verifies that a body which only deletes does not create code that can
  panic; prior to this diff, we would mark the name as using an `Anywhere`,
  but we would never actually provide the `Anywhere` in the types table
  which would panic when we used the parameter as a capture

Reviewed By: yangdanny97

Differential Revision: D83911880

fbshipit-source-id: 141de6e5b3b5dd8e4599d4fb24dc9d65349229a2