Add style filtering to sourcedb and prefer pyi by default
Summary:
Up to now, I've ignored style filtering for build system checks, since they're not that important. Now that I'm getting close to dogfooding, it's time to actually implement it.

For this, we just pass `style_filter` through and handle it in each source db. For most source dbs, it doesn't matter. In `MapDB`, I implement it for testing reasons mostly.

For Buck, we fully implement it to prefer the given style filter, if present, falling back to `ModuleStyle::Interface` to return a `.pyi` if we can, and finally falling back to whatever we have if all of the above fail to find something.

Reviewed By: kinto0

Differential Revision: D83386889

fbshipit-source-id: 473e9631bdc9e29c129f00ea7a56b1382a39ea12