Add some tests where pattern matching doesn't narrow well
Summary:
The first bug is that we are failing to negate the base flow when merging
against inexhaustive matches, which in the case of a termination in a branch
can cause us to not narrow a type. Fix is upcoming.

While I was working on that, I noticed a separate gap which is that an exhaustive
pattern match against a class is un-negatable right now. We did this on purpose,
because a lot of class matches are nonexhausive (e.g. `case Foo(x=42)` is non
exhaustive, it doesn't rule out other `Foo` values). But some class patterns
*are* exhaustive, and it's surprising to have them be non-negatable; this is something
we should track in a bug and fix although given the low prevalence of match it's
not super urgent.

Note that the other test uses literals because otherwise the missing exhaustive
class handling makes it hard to see the fix have any impact.

Reviewed By: yangdanny97

Differential Revision: D83979465

fbshipit-source-id: 4df91c63d8f97f3dc277ae9502d7adfb54ced354