Eliminate `implicit_else` variable
Summary:
The `implicit_else` variable is redundant: the only case where we
set `implicit_else` to `false` is when we hit an `else`, but that
always implies that `exhaustive` is `true`.

As a result, it was impossible for the check to fail, the variable
is just wasting space and compute and confusing matters.

Reviewed By: yangdanny97

Differential Revision: D83979466

fbshipit-source-id: ee552db14e1d24225d8cd056f4c6413ffca70edc