Simplify `negate_and_merge_flow`
Summary:
In order to make some changes we need to how narrows are handled
across merging of branches, I want to first eliminate the
`negate_and_merge_flow` logic because it bakes in assumptions about
negation that are rigid and will make it difficult to drop narrows
as eagerly as we ought to.

As a result, I want to inline it. It's a pretty small function already
and only called in two places so this is not a big problem, but it's
also doing useless work making sure that the `if` branch goes before
the `else` branch in merge order. I suspect this is a historical artifact
from before we sorted unions(?), it makes absolutely no difference today
because a merge is order invariant.

Reviewed By: yangdanny97

Differential Revision: D83979468

fbshipit-source-id: d60adc029d37c0b9dab42a2b2e436f8648e9e61d