Negate match ops in `base` for inexhaustive merges
Summary:
I want to do a refactor of flow merging that will greatly improve our handling
of narrows (needed because without intersection types, the results of narrowing
are lossy and can behave badly downstream of a merge).

One thing I noticed as I was cataloging the callsites I need to refactor is that
if/else and match are inconsistent in a way that makes no sense - their inexhaustive
cases should be the same, but we were failing to negate previous ops in match case.

This diff fixes the bug, and in the process will make upcoming refactors simpler
since handling is now more consistent.

Reviewed By: yangdanny97

Differential Revision: D83979461

fbshipit-source-id: a793b3fd3512855689a0b9ff8d1d58d4b066c19d