Fix namespace package finding (#1242)
Summary:
Pull Request resolved: https://github.com/facebook/pyrefly/pull/1242

This diff mostly includes updates to an interface and the affected tests.

This diff stores found namespace results, and returns them  yfrom the end of `find_import_filtered` if no higher-priority `FindResult` is found sooner. This is because results in `search-path`, `typeshed`, and `site-package-path` can override a previously found namespace package if they contain a file or regular package match, and we shouldn't just early return from a single find component if we find a namespace result.

Reviewed By: stroxler

Differential Revision: D83847654

fbshipit-source-id: 2a540754e5a8f9061dff1ce4d1b84675acfb78c1