Push FindResult conversion to ModulePath up a level
Summary:
This diff pushes the conversion of a `FindResult` to a `ModulePath` from within `find_module_components` up to `find_module`.

This is necessary because `NamespacePackage`s will need special handling, since they can be overridden at any point by a higher priority find result later in the module find logic. In the next diff, we check for namespace modules and handle them.

Reviewed By: ndmitchell

Differential Revision: D83847655

fbshipit-source-id: 39200f9cc2fd5e164f5f36aca3a3462ab54a9502