Rearrange narrows, fix a bug in guard handling
Summary:
We need to include the guard expression in the negation, otherwise
we'll negate matches that would be exhaustive without a guard which
is unsound.

While I was at it I also moved the negated previous ops bind up; I can't
be bothered to write a test for this (it's hard to think of one - the match
itself would have to somehow depend on the previou op negation) but I'm very
confident that it is correct, by the time we reach a match pattern we do in
fact know the earlier patterns failed which in principle gives us information.

And even though I can't think of a test case where that would actually matter,
I think it's easier to read the code when it's in the right order.

Reviewed By: yangdanny97

Differential Revision: D83979464

fbshipit-source-id: 7e7855053b21520d00957a8d7ec6fb4b5e505944