Export function decorator callees
Summary:
In pysa, model queries can match on the "fully qualified callee" of decorators. For instance, one could write a model query that finds all functions decorated by function `my_module.my_decorator`.

To support this feature, we need pyrefly to export the set of callees for each function decorator. This is done in this diff.

Reviewed By: tianhan0

Differential Revision: D83980532

fbshipit-source-id: 079b6f2dab41855e2862dc633ffac293ca1c0cd1