Export class decorator callees
Summary:
In pysa, model queries can match on the "fully qualified callee" of decorators. For instance, one could write a model query that finds all classes decorated by function my_module.my_decorator.

To support this feature, we need pyrefly to export the set of callees for each class decorator. This is done in this diff.

Reviewed By: stroxler

Differential Revision: D84055880

fbshipit-source-id: e8dca40cf2908f4ec4ae33aacd58e00ba89deb2d