Output sourcedb info for import error message and config dump
Summary:
This diff updates the import error message and dump config text to include details
when a build system is involved. When an origin can be found in the sourcedb (the file performing an import), we can also output information around which target is importing it for debugging purposes.

For dump-config, we output the target for the file we're doing the dump-config for if only one file is passed in for a given target

Reviewed By: grievejia

Differential Revision: D84020774

fbshipit-source-id: 1c9898cd03884885ab6c069a85060b2bb7b7e272