Return class-under-construction when invocation of `__new__` returns a non-explicit `Any` type
Summary:
underconstruction

When constructing a class `C` with `__new__` defined, we attempt to invoke the `__new__(arg)` method and consult the return type of that method to figure out the type of a constructor call `C(arg)`. But when that invocation to `__new__` fails, the return type will become `Any` and the current implementation always make the actual return type take priority, which means `C(arg)` would become `Any` as well.

The "actual return type takes priority over class under construction" behavior is actually dictated by the typing spec (which I found very weird -- to me the more intuitive behavior is to error at the return site of `__new__` and still use the class under construction as the return of the constructor call). But I think there's an argument for not using the actual return type if it's a non-explicit `Any` -- this way, we get more precise typing behavior downstream of the constructor call.

Reviewed By: ndmitchell

Differential Revision: D84017711

fbshipit-source-id: 8b15918c6b9d405e859225f78d5df91ef28aadf8