Add `range` to `Fork`
Summary:
After migrating 3 use cases, I realized it's a lot cleaner to just put this in the
start_fork instead of having the callsites - which tend to be pretty complex already -
deal with an extra local floating around until the end.

Might as well make the change before migrating remaining callsites.

Reviewed By: grievejia

Differential Revision: D84038036

fbshipit-source-id: 3f4df283c1eb668d27c7d61ef0e4bfbcb2bcdd4d