Introduce `Fork` data structure
Summary:
In order to optimize away narrows intelligently for branches, I need more
information than Pyrefly preserves today.

At the moment, we try to minimize clones of Flow extremely aggressively, and as
a result we actually destroy knowledge of the base type. Given that our type
system doesn't model intersections (and as a result, merging narrows can produce
some low-quality types that produce false positives downstream), this is
undesirable. Preserving the base flow in a dedicated data structure will prevent
us from trying to optimize all clones, but the types will help us do a better
job merging flows (and I suspect it will make branching code easier to read
as well - the existing code is swap-heavy, which is efficient but slightly hard
to follow).

In addition to the base flow, following the design sketched in
https://docs.google.com/document/d/1JJgkXJuEb3nKdGj7YY-69qU5zMQsl8mL1uhDpaas-Nk/edit?tab=t.0#heading=h.mcimw73r7l7r,
we want to know when a narrow came from an assert, because asserts
terminate control flow; as a result, in an example like:
```
x: int | str | bytes
if condition():
    assert(isinstance(x, str))
else:
    assert(isinstance(x, bytes))
```
we want to preserve narrows and treat `x` as `str | bytes`, but only
because the narrows come from asserts rather than control flow and
negations.

(In the linked doc, we also discuss the possiblity of a "narrow algebra"
as an alternative to modeling asserts in a special way, but I have not come up
with a great way of making a narrow algebra from what we have - modeling
asserts seems much simpler).

This diff adds the data structure for a `Fork`, where control flow splits
into branches and the flows have to be merged back. The new data structure is
thus far unused, I'll use it to power flow-merging in upcoming diffs.

Reviewed By: grievejia

Differential Revision: D84038040

fbshipit-source-id: 3edbb5578dadda7c7fa51e89d00015f751932543