Cut if/else over to Fork, fix a flow bug
Summary:
This requires two new operations on forks:
- Non-exhaustive merging, used for the case where there is no `else`
  and we aren't able to statically verify that at least one branch is taken;
  we'll wind up using this for matches as well in an upcoming diff.
- Abandoning a branch: we need this in order to analyze the *test* of certain
  unreachable code (code guarded by `TYPE_CHECKING` or sys info, which we
  don not analyze) while skipping that branch in the fork merge.

There's a tiny behavior change here: I fixed a bug where we were evaluating
the `test` in the base flow which could cause walrus operations to be incorrectly
treated as defining a name in the base flow rather than in the branch; as a
result in `test_false_and_walrus` we now can see that `value` is *definitely*
not initialized in the second branch.

Note that our handling of walrus in boolean ops is known to be fishy, so even
though this fixes a minor issue with if/else there are some other problems
which I'm not trying to address in this diff, see #1251 for some details.

Reviewed By: grievejia

Differential Revision: D84038038

fbshipit-source-id: b153b736f3384c1a889ffb9b75bdbcd0f465382d