Optimize `Flow` cloning in the `Fork` logic.
Summary:
This diff makes `Fork` uglier but more efficient by replicating the existing
behavior where we make heavy use of `swap` to avoid one unnecessary Flow
clone.

This results in `Scopes` having an invalid empty flow at times (otherwise,
it's hard to avoid ending with two compies of the base flow, which is where
the one extra clone comes from).

That means we have to decouple finishing a branch from starting a branch; in
order to minimize the likelihood of programming errors here I also use a single
bit in the `Fork` to make sure we don't forget to start a branch.

I considered not making this optimization because it definitely adds complexity,
but given how often we do a merge with only one branch (in which case we save
a factor of 2 in flow clones) I concluded we probably should not regress perf
here.

 ---

Note that the callsites don't have to be as verbose as they are, we can
create a few "compound" operations to make things cleaner by combining starting
a fork and a branch / finishing a branch and starting a new one, which I'll do in
the next diff.

But the core operations have to be decoupled, because that's the only way to
minimize cloning in both (rust) loops like the logic for `if/else` and `match`
and straight-line control flow like this boolean expression example; loops will
need to use the bare operations in order to avoid an extra clone at the bottom
of the last loop whereas straight-line control flow can use the compound
operations.

Reviewed By: grievejia

Differential Revision: D84038031

fbshipit-source-id: ad559c9050eee2d72ef5b3f36aac32148eb01ea0