Track a bug in walrus across tests in branching
Summary:
In the preceeding diff, where I fixed a bug where sys-info-falsy tests
that had walrus would mutate the base flow, which was unsound because
most tests aren't necessarily evaluated.

But there's actually one edge case hidden here (which we were already
getting wrong in most cases, we only happened to get it right in
statically-definitely-false cases), which is that the very first test
does always evaluate, so a walrus there puts a new name in the base
flow.

The true flow logic for if/else is pretty complicated:
- The very first test evaluates in the base flow, and can potentially
  mutate it through walrus operators
- After that, every new test evaluates in a special "test flow"
  that is distinct from the base flow, and every branch actually forks
  off of this "test flow". Why? Because we can only reach a lower-down
  test if a previous test evaluated to false, in which case names
  defined by the previous test expressions do in fact have values.

We're only approximating this; right now every branch forks off of the
base flow (with negated previous ops applied), which works very well
for most things but gets certain walrus edge cases like a walrus in the
very first test wrong.

I'm not convinced this is worth investing in right now, walrus isn't
that common and every example I've seen only uses test walrus in
the following branch body. But I figured I should record the false-positive on
uninitialized local in a test case so we don't forget about it.

Reviewed By: grievejia

Differential Revision: D84038034

fbshipit-source-id: 14eeaa38812dc0b3c088f0aa4a9c3d621c061a61