Introduce compound fork operations
Summary:
They make the boolean op example much cleaner.

The simpler operations are still needed in the public API, they are
required to handle if/else chains, pattern matches, and exception
handlers without an extra clone of Flow.

I opted not to create a compound operation for finishing a branch
and the fork because we're going to need multiple variations of
the merge operation and I don't think it's worth duplicating each of
them into a compound operation; if you forget to finish the last
branch you get a panic with a clear error so I think this is okay.

Reviewed By: grievejia

Differential Revision: D84038039

fbshipit-source-id: 0dde453ffcde65620e952747daddf9cdfe1bfb4a