Cut boolean ops over to using `Fork`
Summary:
Before I can use the new `Fork` data structure to power flow merging,
I first need to cut over use cases.

This diff does the first one: boolean operations. The behavior is exactly
as before, but it accomplished by using `start_fork`, `finish_branch`, and
`finish_exhaustive_fork` rather than  by directly manipulating `Flow`
values in the `ensure_expr` code.

In the process, we begin working on the combinators we need for `Fork`. All
use cases are going to be able to use `start_fork` and `finish_branch`, but
we're going to need at least two flavors of "finish" logic to deal with
exhaustive vs inexhaustive cases; we may need more (e.g. one short-term
workaround for issue #1251 is to make boolean ops use more lax merging
which might require another one; that still needs discussion).

NOTE: it might look a little ugly to call `finish_branch` and then
immediately call `finish_exhaustive_fork`; I was briefly tempted to
combine them. But they are distinct concepts, and all of the statement-level
use cases are actually going to call `finish_branch` inside loops so
combining them just based on visual appearance seems like a bad idea.
In addition, if you look at the TODO on bool ops, I actually expect to be
able to get rid of the negation anyway at which point we won't need the
second `finish_branch` anyway.

NOTE: there's a question of efficiency here. We do one extra clone of Flow
in the new code, because we always have a valid base flow in the `Forks` but
we also maintain `self.scopes.current().flow`. The previous logic was a bit more
fiddly and would use swaps to bounce the `base` flow around in a way that let
us skip one clone at the end of a fork. It's likely worth keeping the optimization
but it makes things uglier; I'll stack that in a separate diff.

Reviewed By: grievejia

Differential Revision: D84038035

fbshipit-source-id: b33947b3432de1306dec5ff924d43e723ef72b0b