Tweak the ordering on `base` flow clone
Summary:
I think it's a little weird and inefficient to negate previous ops directly on
the `base` flow in `if/else` branches; it means we're negating things a
quadratic number of times which seems pretty bad.

It does work semantically because repeatedly negating ever-larger prefixes of a
huge "and" operation gives exactly the same types as just negating the largest
relevant prefix, but it's pretty ugly and I don't want to preserve the behavior
across an upcoming refactor.

Reviewed By: grievejia

Differential Revision: D84038032

fbshipit-source-id: 6582d1fd9883aa3f290a8b27bcc7f79f51424ab7