fix hover display for overloaded generics
Summary:
previously, the `overload` appeared at the end of the previous line, and `def function_name` was duplicated for generic functions in overloads, because generic overload signatures keep their name information when converted to types (while non-generic signatures lose their name).

This diff makes it so that we always keep the name by turning the overload signature into a Type::Function instead of a Type::Callable, which allows us to avoid adding the prefix manually.

closes https://github.com/facebook/pyrefly/issues/1261

Reviewed By: grievejia

Differential Revision: D84075944

fbshipit-source-id: 2dbee13da5be5a676185f69d9ea750be708a4ee8