Bugfix: first part of bool op always evaluates
Summary:
We've always had a bug in our bool op code which is that the first
expression in a flow operation actually always evaluates, so any walrus
operations it defines live in the base flow.

When I changed flow merging to stop treating values inititalized in only
one branch as initialized, this became more relevant because before we
were ignoring the resulting unintialized locals, but now we (in most cases
correctly) complain about them.

This diff fixes our handling of that first boolean operation. We actually
will probably have to go farther and unsoundly accept *all* new names
because there are actually two downstream flows, one when the test passes
and another "general" one and we're not equipped to deal with that right
now, but this partial fix is sound and should land first, before the
unsound approximation.

--------

Note that the change in `named_inside_bool_op` is happening because
`b` is actually annotated as `bool` (we know it's `True` but Pyrefly
always ignores the value on AnnAssign); the previous behavior was a
bug caused by us taking the name *after narrowing*; now the base flow
gets the name before narrowing which is correct.

Reviewed By: grievejia

Differential Revision: D84058655

fbshipit-source-id: 9d954686a086cb8acb9444000aec5ccbbaaa253c