Deduplicate `finish_fork` logic
Summary:
In order to work around #1251, we're going to have to create additional
"flavors" of flow merging, which in practice means implementing the logic
with a variety of flags to control various bits of how the merge happens.

As a result, I think it's important to consolidate the two existing
fork merge functions, which already share some logic.

Reviewed By: grievejia

Differential Revision: D84058656

fbshipit-source-id: 3e75e01ec822f8a93523b4c7ffb64b2702b41df3