Disable uninitialized local check inside boolean ops
Summary:
Because there are actually multiple downstream flows when the operation is
used as a guard of an `if`, we need lax behavior to avoid false positives, we're willing
to accept some false negatives to get it.

See the comment in `FlowStyle::merge`, discussion on
https://github.com/facebook/pyrefly/issues/1251,
and the changed test cases for details.

Fixes https://github.com/facebook/pyrefly/issues/1251

I've marked the two tests created specifically for this fix as bugs to track the
false negatives; in all other tests showing the behavior I've just put a comment noting
it is working as intended.

Reviewed By: grievejia

Differential Revision: D84058654

fbshipit-source-id: e8fd667019c30b4c367c836d53f5dcebc5c82223