Handle statically detectable `while True` loops (mostly)
Summary:
In these loops, only a `break` can reach the end, so after the `orelse` (which is
unreachable assuming our static analysis is correct, but we still want to check
it so that we get IDE features and catch bugs as code is being edited) we should
just merge the breaks to create the new flow.

This means that we're excluding the base flow, any `continue` statements, and
the implicit drop off the end of the loop when creating the downstream flow
which is all correct for a `while True` loop.

Fixes https://github.com/facebook/pyrefly/issues/1255, which is a potentially
high-impact false positive for uninitialized local; it also fixes certain
cases where a `while True` narrows a type; I suspect in some codebases these
patterns might matter (in particular using a `while True` until something becomes
non-None, which could occur in retry loops of infrastructure-y code or in
scientific computing when looking for convergence).

I say (mostly) in the title because, while this handling is pretty good and a huge
improvement, I haven't investigated what the best behavior would be for a while
loop with no breaks; I just left the trunk behavior in place and put in a TODO

Reviewed By: grievejia

Differential Revision: D84072003

fbshipit-source-id: 317e219953de514ba5a04398c139dc982026acad